requirements.txt
Django>=4.0.1,<4.1
django-cors-headers==3.14.0
drf-spectacular>=0.22.1,<0.23
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
Pillow>=9.1.0,<9.2
django-model-utils
django-ckeditor
whitenoise